+++ src/aks-preview/azext_aks_preview/_params.py
@@ -131,7 +131,8 @@ def load_arguments(self, _):
         c.argument('ppg')
         c.argument('workspace_resource_id')
         c.argument('enable_msi_auth_for_monitoring',
-                   arg_type=get_three_state_flag(), is_preview=True)
+                   arg_type=
+                   get_three_state_flag(), is_preview=True)
         c.argument('skip_subnet_role_assignment', action='store_true')
         c.argument('enable_fips_image', action='store_true')
         c.argument('enable_cluster_autoscaler', action='store_true')